подключение hoc
