+++ src/components/movie-overview/movie-overview.jsx
@@ -1,39 +1,10 @@
 import React, {Fragment} from "react";
 import PropTypes from "prop-types";
 import {propsForFilms} from "../../types";
-
-const RatingScore = {
-  BAD: 0,
-  NORMAL: 3,
-  GOOD: 5,
-  VERY_GOOD: 8,
-  AWESOME: 10
-};
+import {getTextScore} from "../../utils";
 
 const MovieOverview = (props) => {
   const {films} = props;
-
-  const getTextScore = (ratingScore) => {
-    let textScore = null;
-    switch (true) {
-      case ratingScore >= RatingScore.BAD && ratingScore < RatingScore.NORMAL:
-        textScore = `Bad`;
-        break;
-      case ratingScore >= RatingScore.NORMAL && ratingScore < RatingScore.GOOD:
-        textScore = `Normal`;
-        break;
-      case ratingScore >= RatingScore.GOOD && ratingScore < RatingScore.VERY_GOOD:
-        textScore = `Good`;
-        break;
-      case ratingScore >= RatingScore.VERY_GOOD && ratingScore < RatingScore.AWESOME:
-        textScore = `Very Good`;
-        break;
-      case ratingScore === RatingScore.AWESOME:
-        textScore = `Awesome`;
-        break;
-    }
-    return textScore;
-  };
 
   return (
     <Fragment>
